backend/requirements.txt
# ── Web framework ─────────────────────────────────────────────────────
fastapi==0.111.0
uvicorn[standard]==0.29.0
python-multipart==0.0.9

# ── Google ADK + Gemini ───────────────────────────────────────────────
google-adk==1.0.0
google-generativeai>=0.7.0

# ── GitHub HTTP client ────────────────────────────────────────────────
httpx==0.27.0

# ── Database ──────────────────────────────────────────────────────────
aiosqlite==0.20.0

# ── Auth ──────────────────────────────────────────────────────────────
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
cryptography==42.0.5

# ── Validation ────────────────────────────────────────────────────────
pydantic[email]==2.7.1

# ── Env ───────────────────────────────────────────────────────────────
python-dotenv==1.0.1
